Update Travis job
This upgrades the Travis job to use `cabal new-build`, which has
built-in logic for building and caching local packages such as using an
unreleased `semigroupoids` Git snapshot.

The .cabal file is updated to cabal 1.10 format, and other-extensions
are declared.

Due to the import of GHC.Generics, only GHC 7.4 or later is supported.